+++ src/main/java/recsys/movielens/similarity/SetSimilarity.java
@@ -4,6 +4,6 @@ import java.util.Set;
 
 public interface SetSimilarity {
 
-	public Float getSimilarity(Set<String> set1, Set<String> set2);
+	public Double getSimilarity(Set<String> set1, Set<String> set2);
 
 }
